feat: enhance chat functionality and AI integration
- Add streaming support and "fast" mode for Gemini API
- Implement message deletion and visibility filtering
- Fix SQLCipher passphrase conversion and RTL icons
- Add Melipayamak configuration placeholders

+++ app/src/main/java/com/example/data/database/AppDatabase.kt
@@ -29,7 +29,7 @@ abstract class AppDatabase : RoomDatabase() {
                     try {
                         val db = SQLiteDatabase.openDatabase(
                             dbFile.absolutePath,
-                            passphrase.map { it.toChar() }.toCharArray(),
+                            passphrase.map { it.toInt().toChar() }.toCharArray(),
                             null,
                             SQLiteDatabase.OPEN_READWRITE
                         )
